refactor: Renamed Chrono.Common to Chrono.Shared
