+++ src/Chrono/ConfigureSecurity.cs
@@ -1,5 +1,5 @@
 using System.Security.Claims;
-using Chrono.Common.Interfaces;
+using Chrono.Shared.Interfaces;
 using Chrono.Entities;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
