+++ src/Chrono/Features/Users/GetUserSettings.cs
@@ -1,7 +1,7 @@
-﻿using Chrono.Common.Api;
-using Chrono.Common.Exceptions;
-using Chrono.Common.Interfaces;
-using Chrono.Common.Services;
+﻿using Chrono.Shared.Api;
+using Chrono.Shared.Exceptions;
+using Chrono.Shared.Interfaces;
+using Chrono.Shared.Services;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
