+++ src/Chrono/Infrastructure/ConfigureServices.cs
@@ -1,5 +1,5 @@
-using Chrono.Common.Audit;
-using Chrono.Common.Interfaces;
+using Chrono.Shared.Audit;
+using Chrono.Shared.Interfaces;
 using Chrono.Features.TaskLists;
 using Chrono.Features.Tasks;
 using Chrono.Infrastructure.Persistence;
