+++ src/Chrono/Shared/Behaviors/ValidationBehavior.cs
@@ -1,7 +1,7 @@
 using FluentValidation;
 using MediatR;
 
-namespace Chrono.Common.Behaviors;
+namespace Chrono.Shared.Behaviors;
 
 public class ValidationBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
     where TRequest : notnull
