+++ src/Chrono/Shared/Exceptions/NotFoundException.cs
@@ -1,4 +1,4 @@
-namespace Chrono.Common.Exceptions;
+namespace Chrono.Shared.Exceptions;
 
 public class NotFoundException : Exception
 {
